refactor: update import path for getBlockLineAndOffset test module
